# apps/docs-site/public/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Claustrum Memory Core API",
    "version": "0.1.3",
    "description": "Auto-generated from memory-core route declarations and inline Zod validations."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "API Key"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {}
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/v1/oidc/providers": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "GET /v1/oidc/providers",
        "operationId": "get_v1_oidc_providers",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "workspace_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-oidc-outbound-routes.ts"
      },
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "POST /v1/oidc/providers",
        "operationId": "post_v1_oidc_providers",
        "responses": {
          "201": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-oidc-outbound-routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspace_key": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "issuer_url": {
                    "type": "string"
                  },
                  "client_id": {
                    "type": "string"
                  },
                  "client_secret": {
                    "type": "string"
                  },
                  "discovery_enabled": {
                    "type": "boolean"
                  },
                  "scopes": {
                    "type": "string"
                  },
                  "claim_groups_name": {
                    "type": "string"
                  },
                  "claim_groups_format": {
                    "type": "string",
                    "enum": [
                      "id",
                      "name"
                    ]
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "workspace_key",
                  "name",
                  "issuer_url",
                  "client_id",
                  "client_secret"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/oidc/providers/{id}": {
      "patch": {
        "tags": [
          "Settings"
        ],
        "summary": "PATCH /v1/oidc/providers/:id",
        "operationId": "patch_v1_oidc_providers_id",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-oidc-outbound-routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspace_key": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "issuer_url": {
                    "type": "string"
                  },
                  "client_id": {
                    "type": "string"
                  },
                  "client_secret": {
                    "type": "string"
                  },
                  "discovery_enabled": {
                    "type": "boolean"
                  },
                  "scopes": {
                    "type": "string"
                  },
                  "claim_groups_name": {
                    "type": "string"
                  },
                  "claim_groups_format": {
                    "type": "string",
                    "enum": [
                      "id",
                      "name"
                    ]
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "workspace_key"
                ]
              }
            }
          }
        }
      }
    },
...
    "/v1/outbound-policies/{integration_type}": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "GET /v1/outbound-policies/:integration_type",
        "operationId": "get_v1_outbound_policies_integration_type",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "integration_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "slack",
                "jira",
                "confluence",
                "notion",
                "webhook",
                "email"
              ]
            }
          },
          {
            "name": "workspace_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-oidc-outbound-routes.ts"
      },
      "put": {
        "tags": [
          "Settings"
        ],
        "summary": "PUT /v1/outbound-policies/:integration_type",
        "operationId": "put_v1_outbound_policies_integration_type",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "integration_type",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "slack",
                "jira",
                "confluence",
                "notion",
                "webhook",
                "email"
              ]
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-oidc-outbound-routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspace_key": {
                    "type": "string"
                  },
                  "enabled": {
                    "type": "boolean"
                  },
                  "locale_default": {
                    "type": "string",
                    "enum": [
                      "en",
                      "ko",
                      "ja",
                      "es",
                      "zh"
                    ]
                  },
                  "supported_locales": {
                    "type": "string",
                    "enum": [
                      "en",
                      "ko",
                      "ja",
                      "es",
                      "zh"
                    ]
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "template",
                      "llm"
                    ]
                  },
                  "style": {
                    "type": "string",
                    "enum": [
                      "short",
                      "normal",
                      "verbose"
                    ]
                  },
                  "template_overrides": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "llm_prompt_system": {
                    "type": "string",
                    "nullable": true
                  },
                  "llm_prompt_user": {
                    "type": "string",
                    "nullable": true
                  },
                  "reason": {
                    "type": "string"
                  }
                },
                "required": [
                  "workspace_key"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/outbound/render": {
      "post": {
        "tags": [
          "Settings"
        ],
        "summary": "POST /v1/outbound/render",
        "operationId": "post_v1_outbound_render",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-oidc-outbound-routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "workspace_key": {
                    "type": "string"
                  },
                  "integration_type": {
                    "type": "string",
                    "enum": [
                      "slack",
                      "jira",
                      "confluence",
                      "notion",
                      "webhook",
                      "email"
                    ]
                  },
                  "action_key": {
                    "type": "string"
                  },
                  "params": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "locale": {
                    "type": "string",
                    "enum": [
                      "en",
                      "ko",
                      "ja",
                      "es",
                      "zh"
                    ]
                  }
                },
                "required": [
                  "workspace_key",
                  "integration_type",
                  "action_key"
                ]
              }
            }
          }
        }
      }
    },
    "/v1/outbound/template-variables": {
      "get": {
        "tags": [
          "Settings"
        ],
        "summary": "GET /v1/outbound/template-variables",
        "operationId": "get_v1_outbound_template_variables",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "workspace_key",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "integration_type",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "slack",
                "jira",
                "confluence",
                "notion",
                "webhook",
                "email"
              ]
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-oidc-outbound-routes.ts"
      }
    },
...
    "/v1/workspaces/{key}/outbound-settings": {
      "get": {
        "tags": [
          "Workspace"
        ],
        "summary": "GET /v1/workspaces/:key/outbound-settings",
        "operationId": "get_v1_workspaces_key_outbound_settings",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-oidc-outbound-routes.ts"
      },
      "put": {
        "tags": [
          "Workspace"
        ],
        "summary": "PUT /v1/workspaces/:key/outbound-settings",
        "operationId": "put_v1_workspaces_key_outbound_settings",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-oidc-outbound-routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "default_outbound_locale": {
                    "type": "string",
                    "enum": [
                      "en",
                      "ko",
                      "ja",
                      "es",
                      "zh"
                    ]
                  },
                  "supported_outbound_locales": {
                    "type": "string",
                    "enum": [
                      "en",
                      "ko",
                      "ja",
                      "es",
                      "zh"
                    ]
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/workspaces/{key}/sso-settings": {
      "get": {
        "tags": [
          "Workspace"
        ],
        "summary": "GET /v1/workspaces/:key/sso-settings",
        "operationId": "get_v1_workspaces_key_sso_settings",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-oidc-outbound-routes.ts"
      },
      "put": {
        "tags": [
          "Workspace"
        ],
        "summary": "PUT /v1/workspaces/:key/sso-settings",
        "operationId": "put_v1_workspaces_key_sso_settings",
        "responses": {
          "200": {
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "x-source-file": "apps/memory-core/src/http/routes/core-settings-oidc-outbound-routes.ts",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "oidc_sync_mode": {
                    "type": "string",
                    "enum": [
                      "add_only",
                      "add_and_remove"
                    ]
                  },
                  "oidc_allow_auto_provision": {
                    "type": "boolean"
                  },
                  "reason": {
                    "type": "string"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
...
}